pyproject.toml
[project]
name = "pie-tricotissage-2024-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scipy>=1.15.2",
]
